Options for hash, keysize, altnames, and to make expired certs
Added the ability to set the md hash type ( sha256/sha1 )
Added the ability to set the keysize ( 1024/2048/... )
Added the ability to create either a valid cert date validity, or one that has already expired, for testing purposes.
Added the ability to specify altnames.
Extended the initial example in redo to have a longer chain.
Made it so that ca conf files are generated.
Added various useful comments to the code so that it can be useful for other purposes beyond what it has been made for; adaptable into a general purpose script with some more work.